Create

+++ app/src/main/java/com/example/testtask/data/mappers/imagemappers/ImageToEntityMapper.kt
@@ -12,7 +12,7 @@ class ImageToEntityMapper @Inject constructor() {
         return ImageEntity(
             id = id,
             url = url,
-            date = date.toLong(),
+            date = date,
             lat = lat,
             lng = lng
         )
